+++ src/main/java/io/xpring/xrpl/XRPClient.java
@@ -2,8 +2,6 @@ package io.xpring.xrpl;
 
 import java.math.BigInteger;
 
-import io.xpring.xrpl.legacy.LegacyDefaultXRPClient;
-
 /**
  * A client that can submit transactions to the XRP Ledger.
  *
@@ -13,27 +11,12 @@ public class XRPClient implements XRPClientInterface {
     private XRPClientDecorator decoratedClient;
 
     /**
-     * Initialize a new client.
-     *
-     * The client will use the rippled implementation of protocol buffers.
+     * Initialize a new client with the given options.
      *
      * @param grpcURL The remote URL to use for gRPC calls.
      */
     public XRPClient(String grpcURL) {
-        this(grpcURL, true);
-    }
-
-    /**
-     * Initialize a new client with the given options.
-     *
-     * @param useNewProtocolBuffers:  If `true`, then the new protocol buffer implementation from rippled will be used.
-     *
-     * @param grpcURL The remote URL to use for gRPC calls.
-     */
-    public XRPClient(String grpcURL, boolean useNewProtocolBuffers) {
-        XRPClientDecorator defaultXRPClient = useNewProtocolBuffers ?
-                new DefaultXRPClient(grpcURL) :
-                new LegacyDefaultXRPClient(grpcURL);
+        XRPClientDecorator defaultXRPClient =  new DefaultXRPClient(grpcURL);
         this.decoratedClient = new ReliableSubmissionXRPClient(defaultXRPClient);
     }
 
@@ -92,5 +75,15 @@ public class XRPClient implements XRPClientInterface {
             final Wallet sourceWallet
     ) throws XpringException {
         return decoratedClient.send(amount, destinationAddress, sourceWallet);
+    }
+
+    /**
+     * Check if an address exists on the XRP Ledger.
+     *
+     * @param xrplAccountAddress The address to check the existence of.
+     * @return A boolean if the account is on the XRP Ledger.
+     */
+    public boolean accountExists(final String xrplAccountAddress) throws XpringException {
+        return decoratedClient.accountExists(xrplAccountAddress);
     }
 }