+++ src/test/java/io/xpring/xrpl/DefaultXRPClientTest.java
@@ -16,6 +16,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.AdditionalAnswers.delegatesTo;
 import io.grpc.ManagedChannel;
 import io.grpc.inprocess.InProcessChannelBuilder;
+import io.grpc.StatusRuntimeException;
+import io.grpc.Status;
 import org.xrpl.rpc.v1.*;
 import org.xrpl.rpc.v1.Common.*;
 import java.io.IOException;
@@ -355,6 +357,67 @@ public class DefaultXRPClientTest {
         // WHEN the transactionHistory is requested THEN a conversion error is thrown.
         expectedException.expect(XpringException.class);
         xrpClient.paymentHistory(XRPL_ADDRESS);
+    }
+
+    @Test
+    public void accountExistsTest() throws IOException, XpringException {
+        // GIVEN a DefaultXRPClient with mocked networking which will succeed.
+        DefaultXRPClient client = getClient();
+
+        // WHEN the account is checked
+        boolean exists = client.accountExists(XRPL_ADDRESS);
+
+        // THEN the existence of the account is the the same as the mocked response.
+        assertThat(exists).isEqualTo(true);
+    }
+
+    @Test
+    public void accountExistsWithClassicAddressTest() throws IOException, XpringException {
+        // GIVEN a classic address.
+        ClassicAddress classicAddress = Utils.decodeXAddress(XRPL_ADDRESS);
+        DefaultXRPClient client = getClient();
+
+        // WHEN the existence of the account is checked for the classic address THEN an error is thrown.
+        expectedException.expect(XpringException.class);
+        client.accountExists(classicAddress.address());
+    }
+
+    @Test
+    public void accountExistsTestWithNotFoundError() throws IOException, XpringException {
+        // GIVEN a XRPClient with mocked networking which will fail to retrieve account info w/ NOT_FOUND error code.
+        StatusRuntimeException notFoundError = new StatusRuntimeException(Status.NOT_FOUND);
+        GRPCResult<GetAccountInfoResponse> accountInfoResult = GRPCResult.error(notFoundError);
+        DefaultXRPClient client = getClient(
+                accountInfoResult,
+                GRPCResult.ok(makeTransactionStatus(true, TRANSACTION_STATUS_SUCCESS)),
+                GRPCResult.ok(makeGetFeeResponse(MINIMUM_FEE, LAST_LEDGER_SEQUENCE)),
+                GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH)),
+                GRPCResult.ok(makeGetAccountTransactionHistoryResponse())
+        );
+
+        // WHEN the existence of the account is checked
+        boolean exists = client.accountExists(XRPL_ADDRESS);
+
+        // THEN false is returned.
+        assertThat(exists).isEqualTo(false);
+    }
+
+    @Test
+    public void accountExistsTestWithUnkonwnError() throws IOException, XpringException {
+        // GIVEN a XpringClient with mocked networking which will fail to retrieve account info w/ UNKNOWN error code.
+        StatusRuntimeException notFoundError = new StatusRuntimeException(Status.UNKNOWN);
+        GRPCResult<GetAccountInfoResponse> accountInfoResult = GRPCResult.error(notFoundError);
+        DefaultXRPClient client = getClient(
+                accountInfoResult,
+                GRPCResult.ok(makeTransactionStatus(true, TRANSACTION_STATUS_SUCCESS)),
+                GRPCResult.ok(makeGetFeeResponse(MINIMUM_FEE, LAST_LEDGER_SEQUENCE)),
+                GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH)),
+                GRPCResult.ok(makeGetAccountTransactionHistoryResponse())
+        );
+
+        // WHEN the existence of the account is checked THEN the error is re-thrown.
+        expectedException.expect(StatusRuntimeException.class);
+        client.getBalance(XRPL_ADDRESS);
     }
 
     /**